renamed populateClasses management file to manageclasses, added more arguments
